third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-contain/container-queries/font-relative-units-expected.txt

This is a testharness.js-based test.
[PASS] em relative inline-size
[PASS] rem relative inline-size
[PASS] ex relative inline-size
[FAIL] rex relative inline-size
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
[PASS] ch relative inline-size
[PASS] rch relative inline-size
[PASS] ic relative inline-size
[PASS] ric relative inline-size
[PASS] lh relative inline-size
[PASS] rlh relative inline-size
[PASS] cap relative inline-size
[FAIL] rcap relative inline-size
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
